Create es_ES.php
Hello there,
my team and I decided about a month ago to introduce Cockpit CMS in our architecture. So far so good :-)
And because the company we work for, is active on the Spanish market, the Spanish version of the administration area was/is required.
Therefore based on the Cockpit's document (https://getcockpit.com/documentation/reference/language) we translated the default (English) version into Spanish.
We hope it will help others out there.

Greetings Patrick.